Add relationship between comments and users
Added relationship between comments and users where users have many
comments and comments belong to users. Also added the user parameters
where a user has to be logged in to make a comment and when they make
a comment their user ID is automatically associated with the comment.
Next will be to update the serializers for the resources I've created
and to add the relationships between comments and blogs