Add admin attribute to users
Added admin attribute to users which is a boolean. In the migration,
the default value is set to false. In examples.rb I am creating an user
with an admin value of true. There is no way to create a user with an
admin value of true on the front end.